resuelto reserva de un libro.

+++ src/main/java/edu/cesur/fullstack/model/ReservaDTO.java
@@ -9,11 +9,11 @@ import lombok.Data;
 @AllArgsConstructor
 public class ReservaDTO { 
 
-    private Long id; 
+    private Integer id; 
 
-    private Long libroId; 
+    private Integer libroId; 
 
-    private Long usuarioId; 
+    private Integer usuarioId; 
 
     private LocalDate fechaReserva;
