+++ src/main/java/edu/cesur/fullstack/services/BookServiceImpl.java
@@ -6,27 +6,26 @@ import java.util.Arrays;
 import org.springframework.stereotype.Service;
 
 import edu.cesur.fullstack.model.BookDTO;
-import edu.cesur.fullstack.model.UserDTO;
 
 
 @Service
 public class BookServiceImpl implements BookService {
 	
 	ArrayList<BookDTO> books = new ArrayList(
-			Arrays.asList(new BookDTO(8l, "Aventuras", "Javier", false),
-					new BookDTO(1l, "Viajes locos", "Acner", false),
-					new BookDTO(2l, "Estudiando Full Stack", "José", false),
-					new BookDTO(3l, "Cómo aprobar sin morir", "Marcos", false),
-					new BookDTO(4l, "El conejo no se ha perdido", "Miriam", false),
-					new BookDTO(5l, "El principito", "Andrea", false),
-					new BookDTO(6l, "Harry Potter", "Gian", false),
-					new BookDTO(7l, "Locas aventuras", "Mireia", false)			
+			Arrays.asList(new BookDTO(8, "Aventuras", "Javier", false),
+					new BookDTO(1, "Viajes locos", "Acner", false),
+					new BookDTO(2, "Estudiando Full Stack", "José", false),
+					new BookDTO(3, "Cómo aprobar sin morir", "Marcos", false),
+					new BookDTO(4, "El conejo no se ha perdido", "Miriam", false),
+					new BookDTO(5, "El principito", "Andrea", false),
+					new BookDTO(6, "Harry Potter", "Gian", false),
+					new BookDTO(7, "Locas aventuras", "Mireia", false)			
 			));
 	
 	
 
 	@Override
-	public boolean getStatus(Long id) {
+	public boolean getStatus(Integer id) {
 		
 		BookDTO b = getBookById(id);
 		return b.getReservado();
@@ -35,7 +34,7 @@ public class BookServiceImpl implements BookService {
 	
 
 	@Override
-	public BookDTO getBookById(Long id) {
+	public BookDTO getBookById(Integer id) {
 		for (BookDTO book : books) {
 			if (book.getId() == id) {
 				return book;
